+++ src/include/fw_config.h
@@ -18,8 +18,8 @@
 struct fw_config {
 	const char *field_name;
 	const char *option_name;
-	uint32_t mask;
-	uint32_t value;
+	uint64_t mask;
+	uint64_t value;
 };
 
 /* Generate a pointer to a compound literal of the fw_config structure. */
@@ -53,7 +53,7 @@ void fw_config_for_each_found(void (*cb)(const struct fw_config *config, void *a
  *
  * Return pointer to cached `struct fw_config` if successfully probed, otherwise NULL.
 */
-const struct fw_config *fw_config_get_found(uint32_t field_mask);
+const struct fw_config *fw_config_get_found(uint64_t field_mask);
 
 #else
