+++ src/mainboard/google/dedede/board_info.c
@@ -3,7 +3,7 @@
 #include <baseboard/variants.h>
 #include <ec/google/chromeec/ec.h>
 
-int board_info_get_fw_config(uint32_t *fw_config)
+int board_info_get_fw_config(uint64_t *fw_config)
 {
 	return google_chromeec_cbi_get_fw_config(fw_config);
 }
